create error class

+++ src/storage/group/groupCreate.ts
@@ -1,10 +1,18 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import { GROUP_COLLECTION } from "@storage/storageConfig";
 import { groupsGetAll } from "./groupsGetAll";
+import { AppError } from "@utils/AppError";
 
 export async function groupCreate(newGroup: string) {
     try {
-        const storagedGroups = await groupsGetAll();
+        const storagedGroups = await groupsGetAll()
+
+        const groupAlreadyExists = storagedGroups.includes(newGroup);
+
+        if (groupAlreadyExists) {
+            throw new AppError("Grupo já cadastrado.")
+        }
+
         const storage = JSON.stringify([...storagedGroups, newGroup])
 
         await AsyncStorage.setItem(GROUP_COLLECTION, storage)
